Feat: 결과 페이지 내용 및 관련 컴포넌트 이동 #52
