+++ src/app/(tests)/freetest/[id]/components/MbtiSection/MbtiSection.tsx
