+++ src/app/(tests)/freetest/[id]/components/MbtiSection/_MbtiSection.scss
