+++ src/app/(tests)/freetest/[id]/components/ShareSection/_ShareSection.scss
